added function to restrict access excepting user and to login through facebook

+++ routes/page.js
@@ -6,57 +6,64 @@ const db = require('../lib/db.js');
 
 
 router.get('/toDoList', function(request, response) {
-    var title = "To Do List";
-    var statusUI = auth.statusUI(request, response);
-    var description = '<ul>';
-    var achievement_rate = '';
-    var control = '';
-    var list = db.get('toDoList').value();
-    if(1 <= list.length) {
-        for(var i = 0 ; i < list.length ; i++) {
-            description = description + `<li>${list[i].title}</li>`
-        }
-        description += '</ul>';
-        achievement_rate = `<div id="rate">
-                                <div id="rate_text"><div><br><br>달성율(%) : </div><div id="graph_num">0</div></div>
-                                <div id="graph"><div id="progress">|</div></div>
-                            </div>`;
-
-        
-        control = `
-        <div class="order"><input type="button" value="수정하기" onclick="updateToDoList()"> <input type="button" value="모두 삭제" onclick="deleteToDoList()"></div>`
-        console.log(list);
+    if(request.user == undefined) {
+        response.redirect('/auth/login');
     } else {
-        description = `<p>등록된 할 일이 없습니다.</p>`
-        control = `<div class="order"><input type="button" value="오늘의 할일 설정하기" onclick="createToDoList()"></div>`
-    }
-
-    var html = template.HTML(title, description, achievement_rate, control, statusUI);
-    response.send(html);
-});
+        var title = "To Do List";
+        var statusUI = auth.statusUI(request, response);
+        var description = '<ul>';
+        var achievement_rate = '';
+        var control = '';
+        var list = db.get('toDoList').filter({user_id:request.user.shortid}).value();
+        console.log("list-ft : ", list);
+        if(1 <= list.length) {
+            for(var i = 0 ; i < list.length ; i++) {
+                description = description + `<li>${list[i].title}</li>`
+            }
+            description += '</ul>';
+            achievement_rate = `<div id="rate">
+                                    <div id="rate_text"><div><br><br>달성율(%) : </div><div id="graph_num">0</div></div>
+                                    <div id="graph"><div id="progress">|</div></div>
+                                </div>`;
+            control = `
+            <div class="order"><input type="button" value="수정하기" onclick="updateToDoList()"> <input type="button" value="모두 삭제" onclick="deleteToDoList()"></div>`
+            console.log(list);
+        } else {
+            description = `<p>등록된 할 일이 없습니다.</p>`
+            control = `<div class="order"><input type="button" value="오늘의 할일 설정하기" onclick="createToDoList()"></div>`
+        }
+    
+        var html = template.HTML(title, description, achievement_rate, control, statusUI);
+        response.send(html);
+        }
+    });
 
 
 router.get('/create_ToDoList', function(request, response) {
-    var title = "Create To Do List";
-    var statusUI = auth.statusUI(request, response);
-    var html = template.HTML(title, 
-    `<form action="/page/create_ToDoList_process" method="post">
-        <div id="new_list">
-            <div></div>
-            <div id="first_list">
-                <input type="text" class="list" name="input" placeholder="오늘의 할일 1">
-                <div id="added_list"></div>
-            </div>
-            <div>
-                <input type="button" value="항목 추가" onclick="add_List()"> 
-                <input type="button" value="항목 삭제" onclick="delete_List()">
-            </div>
-            <div id="listCount"><input type="hidden" name="listCount" value=1></div>
-        </div>
-        <p id="submit"><input type="submit" value="완료"></p>
-    </form>
-    `, '', '', statusUI);
-    response.send(html);
+    if(request.user == undefined) {
+        response.redirect('/auth/login');
+    } else {
+        var title = "Create To Do List";
+        var statusUI = auth.statusUI(request, response);
+        var html = template.HTML(title, 
+            `<form action="/page/create_ToDoList_process" method="post">
+                <div id="new_list">
+                    <div></div>
+                    <div id="first_list">
+                        <input type="text" class="list" name="input" placeholder="오늘의 할일 1">
+                        <div id="added_list"></div>
+                    </div>
+                    <div>
+                        <input type="button" value="항목 추가" onclick="add_List()"> 
+                        <input type="button" value="항목 삭제" onclick="delete_List()">
+                    </div>
+                    <div id="listCount"><input type="hidden" name="listCount" value=1></div>
+                </div>
+                <p id="submit"><input type="submit" value="완료"></p>
+            </form>
+            `, '', '', statusUI);
+        response.send(html);
+    }
 });
 
 
@@ -75,7 +82,8 @@ router.post("/create_ToDoList_process", function(request, response) {
         if(list[i] != "") {
             db.get('toDoList').push({
                 title: list[i],
-                num: listNum
+                num: listNum,
+                user_id:request.user.shortid
             }).write();
             listNum++;
         }  
@@ -84,41 +92,45 @@ router.post("/create_ToDoList_process", function(request, response) {
     response.redirect('/page/toDoList');
 });
 
-router.get("/update_ToDoList", function(request, response) {
-    var list = db.get('toDoList').value();
-    var title = "To Do List - 수정"; 
-    var description = "";
-    var statusUI = auth.statusUI(request, response);
+router.get("/update_ToDoList", function(request, response) {  
+    if(request.user == undefined) {
+        response.redirect('/auth/login');
+    } else {
+        var list = db.get('toDoList').filter({user_id:request.user.shortid}).value();
+        var title = "To Do List - 수정"; 
+        var description = "";
+        var statusUI = auth.statusUI(request, response);
 
-    for(var i = 0; i < list.length; i++) {
-        description += `
-        <div id='list_${i+1}'><p><input type='text' name='input' class='list' placeholder='오늘의 할일${i+1}' value="${list[i].title}"></p></div>`;
-    }
-    
-    var html = template.HTML(title,
-    `
-    <form action="/page/update_ToDoList_process" method="post">
-        <div id="new_list">
-            <div></div>
-            <div id="added_list">${description}</div>
-            <div id="controler">
-                <input type="button" value="항목 추가" onclick="add_List()"> 
-                <input type="button" value="항목 삭제" onclick="delete_List()">
+        for(var i = 0; i < list.length; i++) {
+            description += `
+            <div id='list_${i+1}'><p><input type='text' name='input' class='list' placeholder='오늘의 할일${i+1}' value="${list[i].title}"></p></div>`;
+        }
+        
+        var html = template.HTML(title,
+        `
+        <form action="/page/update_ToDoList_process" method="post">
+            <div id="new_list">
+                <div></div>
+                <div id="added_list">${description}</div>
+                <div id="controler">
+                    <input type="button" value="항목 추가" onclick="add_List()"> 
+                    <input type="button" value="항목 삭제" onclick="delete_List()">
+                </div>
+                <div id="listCount"><input type="hidden" name="listCount" value=${list.length}></div>
             </div>
-            <div id="listCount"><input type="hidden" name="listCount" value=${list.length}></div>
-        </div>
-        <p id="submit"><input type="submit" value="완료"></p>
-    </form>
-    `, ``, ``, statusUI, list.length);
-
-    response.send(html);
+            <p id="submit"><input type="submit" value="완료"></p>
+        </form>
+        `, ``, ``, statusUI, list.length);
+    
+        response.send(html);
+    }
 })
 
 router.post('/update_ToDoList_process', function(request, response) {
-    var list = db.get('toDoList').value();
+    var list = db.get('toDoList').filter({user_id:request.user.shortid}).value();
     var listcount = list.length;
     for(var i = 0 ; i < listcount ; i++) {
-        db.get('toDoList').remove({num:i}).write();
+        db.get('toDoList').remove({user_id:request.user.shortid, num:i}).write();
     }
     var post = request.body;
     var listCount = post.listCount;
@@ -134,7 +146,8 @@ router.post('/update_ToDoList_process', function(request, response) {
         if(list[i] != "") {
             db.get('toDoList').push({
                 title: list[i],
-                num: listNum
+                num: listNum,
+                user_id: request.user.shortid
             }).write();
             listNum++;
         }  
@@ -143,12 +156,16 @@ router.post('/update_ToDoList_process', function(request, response) {
 })
 
 router.get('/delete_toDoList_process', function(request, response) {
-    var list = db.get('toDoList').value();
-    var listcount = list.length;
-    for(var i = 0 ; i < listcount ; i++) {
-        db.get('toDoList').remove({num:i}).write();
+    if(request.user == undefined) {
+        response.redirect('/auth/login');
+    } else {
+        var list = db.get('toDoList').filter({user_id:request.user.shortid}).value();
+        var listcount = list.length;
+        for(var i = 0 ; i < listcount ; i++) {
+            db.get('toDoList').remove({user_id:request.user.shortid, num:i}).write();
+        }
+        response.redirect('/page/toDoList');
     }
-    response.redirect('/page/toDoList');
 })
